requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
et_xmlfile==2.0.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==2.1.3
openai==0.28.0
openpyxl==3.1.5
pandas==2.2.3
pandas-stubs==2.2.3.241126
propcache==0.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
tqdm==4.67.1
types-pytz==2024.2.0.20241003
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.3
